Read comment please
this is what the df.head looks like

How can i download it from an external server or export it in a way so i can use the data. 
I have displayed the whole data frame as a string but couldn't safe it as .csv data (the formatting was all in now field).

Iam a newbe and need help for this. Thank you guys a lot for any suggestions.